Move get_plugin_manager into conda.plugins
