+++ tests/plugins/conftest.py
@@ -12,10 +12,13 @@ from conda import plugins
 
 @pytest.fixture
 def plugin_manager(mocker):
-    pm = pluggy.PluginManager('conda')
-    pm.add_hookspecs(plugins)
-    mocker.patch('conda.base.context.get_plugin_manager', return_value=pm)
-    return pm
+    manager = pluggy.PluginManager("conda")
+    manager.add_hookspecs(plugins.specs)
+    mocker.patch(
+        "conda.plugins.manager",
+        manager,
+    )
+    return manager
 
 
 @pytest.fixture
